chore: Created subproject for cost report of infrastructures
